Added routes to OwnerService
