+++ app/owner/__init__.py
@@ -2,4 +2,4 @@ from flask import Blueprint
 
 owner = Blueprint('owner', __name__)
 
-from . import views, errors
+from . import OwnerService, views, errors
