+++ app/owner/views.py
@@ -9,19 +9,3 @@ from datetime import date
 @owner.route('/')
 def index():
     return render_template('owner/index.html')
-
-@owner.route('/ownerModule/ownerView')
-def ownerView():
-    return render_template('owner/ownerView.html')
-
-@owner.route('/ownerModule/getWorkers')
-def getWorkers():
-    return render_template('owner/getWorkersView.html',
-                           users=User.query.order_by(User.id.desc()).all()
-                           )
-
-@owner.route('/ownerModule/getWorkSchedules')
-def getWorkSchedules():
-    return render_template('owner/workSchedules.html',
-                           workSchedules=WorkSchedule.query.order_by(WorkSchedule.id.desc()).all()
-                           )
